+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## v0.12.0 / 2014-06-19
 
 * use syscall.ERROR_MORE_DATA from ztypes_windows (Go 1.3+)
+* [internal] remove cookie from Event struct (unused)
 
 ## v0.11.0 / 2014-06-12